KP-14175 Added test for base url extraction from IHeadquartersApi
